requirements.txt
requests
python-dotenv
qgenie-sdk
rich
flask
waitress
